+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ### Added
 
+- Add `ToExecuteCommandMatching()` expectation
 - Add `ToRepeatedly()` expectation
 
 ## [0.13.0] - 2021-03-01